Use Authorization header for the API key
